requirements.txt
# Core dependencies
yt-dlp>=2023.12.30
aiohttp>=3.9.0
asyncio-mqtt>=0.13.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# CLI interface
click>=8.1.0
rich>=13.7.0
textual>=0.45.0

# API interface
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-socketio>=5.10.0

# Database
aiosqlite>=0.19.0
asyncpg>=0.29.0
sqlalchemy>=2.0.0

# Queue management
redis>=5.0.0
celery>=5.3.0

# Desktop GUI (optional)
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0

# Multi-engine support
aria2p>=0.11.3
streamlink>=6.0.0
gallery-dl>=1.26.0

# Utilities
pathlib2>=2.3.7
httpx>=0.25.0
python-multipart>=0.0.6
jinja2>=3.1.0
psutil>=5.9.0

# Notifications (optional)
plyer>=2.1.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
